同步 Boot 代码, 较大变动

+++ src/main/java/com/pearadmin/modules/system/service/impl/SysDictDataServiceImpl.java
@@ -2,6 +2,10 @@ package com.pearadmin.modules.system.service.impl;
 
 import com.github.pagehelper.PageHelper;
 import com.github.pagehelper.PageInfo;
+import com.google.common.cache.CacheBuilder;
+import com.google.common.cache.CacheLoader;
+import com.google.common.cache.LoadingCache;
+import com.pearadmin.common.tools.spring.SpringUtil;
 import com.pearadmin.common.web.domain.request.PageDomain;
 import com.pearadmin.modules.system.domain.SysDictData;
 import com.pearadmin.modules.system.mapper.SysDictDataMapper;
@@ -9,7 +13,10 @@ import com.pearadmin.modules.system.service.ISysDictDataService;
 import org.springframework.stereotype.Service;
 
 import javax.annotation.Resource;
+import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.TimeUnit;
 
 /**
  * Describe: 字典值服务实现类
@@ -21,10 +28,38 @@ public class SysDictDataServiceImpl implements ISysDictDataService {
 
     @Resource
     private SysDictDataMapper sysDictDataMapper;
-
+    //字典缓存 10分钟失效
+    public static LoadingCache<String, List<SysDictData>> loadingCacheSysDictData = CacheBuilder.newBuilder().maximumSize(100).expireAfterWrite(600, TimeUnit.SECONDS).build(new CacheLoader<String,List<SysDictData>>() {
+        @Override
+        public List<SysDictData> load(String typeCode) {
+            SysDictDataMapper tempSysDictDataMapper =SpringUtil.getBean("sysDictDataMapper",SysDictDataMapper.class);
+            return tempSysDictDataMapper.selectByCode(typeCode);
+        }
+    });
     @Override
     public List<SysDictData> list(SysDictData sysDictData) {
         return sysDictDataMapper.selectList(sysDictData);
+    }
+
+    @Override
+    public List<SysDictData> selectByCode(String typeCode) {
+        try {
+            List<SysDictData>  list=  loadingCacheSysDictData.get(typeCode);
+            return list;
+        } catch (ExecutionException e) {
+            e.printStackTrace();
+        }
+        return new ArrayList<>();
+
+    }
+
+    @Override
+    public void refreshChcheTypeCode(String typeCode) {
+        try {
+            loadingCacheSysDictData.refresh(typeCode);
+        }catch (Exception e){
+
+        }
     }
 
     @Override
@@ -38,6 +73,7 @@ public class SysDictDataServiceImpl implements ISysDictDataService {
     public Boolean save(SysDictData sysDictData) {
         Integer result = sysDictDataMapper.insert(sysDictData);
         if(result>0){
+            refreshChcheTypeCode(sysDictData.getTypeCode());
             return true;
         }else{
             return false;
@@ -53,6 +89,7 @@ public class SysDictDataServiceImpl implements ISysDictDataService {
     public Boolean updateById(SysDictData sysDictData) {
         int result = sysDictDataMapper.updateById(sysDictData);
         if(result > 0){
+            refreshChcheTypeCode(sysDictData.getTypeCode());
             return true;
         }else{
             return false;
@@ -61,8 +98,13 @@ public class SysDictDataServiceImpl implements ISysDictDataService {
 
     @Override
     public Boolean remove(String id) {
-        int result = sysDictDataMapper.deleteById(id);
+        SysDictData sysDictData=  sysDictDataMapper.selectById(id);
+        int result=0;
+        if(sysDictData!=null) {
+             result = sysDictDataMapper.deleteById(id);
+        }
         if(result>0){
+            refreshChcheTypeCode(sysDictData.getTypeCode());
             return true;
         }else{
             return false;
